Refactor divisor data type casting

+++ src/main/java/org/candlepin/subscriptions/tally/facts/FactNormalizer.java
@@ -224,7 +224,7 @@ public class FactNormalizer {
         virtualFacts.getSystemProfileCoresPerSocket() * virtualFacts.getSystemProfileSockets();
 
     var threadsPerCore = 2.0;
-    return Double.valueOf(Math.ceil(cpu / threadsPerCore)).intValue();
+    return (int) Math.ceil(cpu / threadsPerCore);
   }
 
   private void getProductsFromProductIds(
